requirements.txt
annotated-types==0.7.0
anyio==4.13.0
catboost==1.2.7
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.115.6
fonttools==4.62.1
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
joblib==1.4.2
kiwisolver==1.5.0
matplotlib==3.10.8
narwhals==2.19.0
numpy==1.26.4
openai==2.31.0
packaging==26.0
pandas==2.2.2
pillow==12.2.0
plotly==6.6.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
scikit-learn==1.5.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.34.0
watchfiles==1.1.1
websockets==16.0
